godream/requirements.txt
django==1.11.11
djangorestframework==3.7.7
virtualenv==13.1.2
wheel==0.24.0
